requirements.txt
fastapi
gunicorn
uvicorn[standard]
torch
transformers
pydantic
pandas
numpy
scikit-learn
datasets
evaluate
matplotlib
azure-storage-blob
